添加移动球体

+++ MyRayTracer/src/Ray.h
@@ -6,7 +6,7 @@ class Ray
 {
 public:
 	Ray() = default;
-	Ray(const vec3& o, const vec3& d, double t)
+	Ray(const vec3& o, const vec3& d, float t)
 		: orig(o), dir(d), tm(t) {}
 	Ray(const vec3& o, const vec3& d)
 		: Ray(o, d, 0) {}
@@ -14,7 +14,7 @@ public:
 	const vec3& GetOrigin() const { return orig; }
 	const vec3& GetDirection() const { return dir; }
 
-	double time() const { return tm; }
+	float time() const { return tm; }
 
 	vec3 at(float t) const {
 		return orig + t * dir;
@@ -23,5 +23,5 @@ public:
 private:
 	vec3 orig;
 	vec3 dir;
-	double tm;
+	float tm;
 };
